+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.6.0
+* Give precedence to column attribute accessors instead of extra_attributes during store_attributes memoization.
+* Update `is_privilege` and `accessible_objects` to use the assignments join table instead of the operations policy elements associations table.
+
 ## 1.5.3
 * Add optional filtering to parents and children attribute accessors in the ActiveRecord storage
   adapter.